Revert "refactors code to commonalise folder and page selectors"
This reverts commit 3517bad5e7ee32804fec3369f61ccd4c0e09733f.
